Made the wallet balance hold satoshis

+++ backend/common/src/main/java/com/kollybistes/common/models/BitcoinWallet.java
@@ -4,10 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.data.annotation.LastModifiedDate;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 @Entity
@@ -20,7 +19,7 @@ public class BitcoinWallet {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String address;
-    private BigDecimal balance;
+    private BigInteger balance;
     private String privateKey;
     private String publicKey;
     @CreationTimestamp
